test: disable schema tests on sqlite

+++ test/integration/include/schema.test.js
@@ -6,7 +6,8 @@ const expect = chai.expect;
 const Support = require('../support');
 const { DataTypes, Op, Sequelize } = require('@sequelize/core');
 
-const dialect = Support.getTestDialect();
+const dialect = Support.sequelize.dialect;
+const dialectName = Support.getTestDialect();
 const _ = require('lodash');
 const promiseProps = require('p-props');
 
@@ -15,6 +16,10 @@ const sortById = function (a, b) {
 };
 
 describe(Support.getTestDialectTeaser('Includes with schemas'), () => {
+  if (!dialect.supports.schemas) {
+    return;
+  }
+
   describe('findAll', () => {
     afterEach(async function () {
       await this.sequelize.dropSchema('account');
@@ -1180,7 +1185,7 @@ describe(Support.getTestDialectTeaser('Includes with schemas'), () => {
         include: [Group],
       });
 
-      if (dialect === 'sqlite') {
+      if (dialectName === 'sqlite') {
         expect(new Date(users[0].dateField).getTime()).to.equal(Date.UTC(2014, 1, 20));
         expect(new Date(users[0].groups[0].dateField).getTime()).to.equal(Date.UTC(2014, 1, 20));
       } else {
